bits and bobs related to world storage

+++ src/game/GlobalBits.java
@@ -10,7 +10,7 @@ import java.util.ArrayList;
 public class GlobalBits {
     public static Render render;
     public static Vector3f playerPosition;
-    public static float renderDistance;
+    public static int renderDistance;
     public static ArrayList<Block> blocks;
     public static String resourcesPath;
     public static final Block voidBlock = new VoidBlock();
